streamlit/requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
